correction prepare for segue et didSelect methods
